+++ app/src/main/java/org/greenstand/android/TreeTracker/activities/ImageCaptureActivity.kt
@@ -20,7 +20,6 @@ class CaptureImageContract : ActivityResultContract<Boolean, String?>() {
 
     companion object {
         const val SELFIE_MODE = "SELFIE_MODE"
-        const val FOCUS_METRIC_VALUE = "FOCUS_METRIC_VALUE"
         const val TAKEN_IMAGE_PATH = "TAKEN_IMAGE_PATH"
     }
 
@@ -41,28 +40,8 @@ class CaptureImageContract : ActivityResultContract<Boolean, String?>() {
 
 class ImageCaptureActivity : AppCompatActivity() {
 
-//    private lateinit var viewFinder: TextureView
-//    private lateinit var imageCaptureButton: ImageButton
-//    private lateinit var toolbarTitle: TextView
-//    private val deviceOrientation by inject<DeviceOrientation>()
-//
-    companion object {
-        private const val SELFIE_MODE = "SELFIE_MODE"
-
-        const val FOCUS_METRIC_VALUE = "FOCUS_METRIC_VALUE"
-        const val TAKEN_IMAGE_PATH = "TAKEN_IMAGE_PATH"
-
-        fun createIntent(context: Context, selfieMode: Boolean = false): Intent {
-            return Intent(context, ImageCaptureActivity::class.java).apply {
-                putExtra(SELFIE_MODE, selfieMode)
-            }
-        }
-    }
-
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
-        val captureSelfie = intent.extras?.getBoolean(SELFIE_MODE, false) ?: false
 
         setContent {
             val navController = rememberNavController()
